Update GitHub actions versions and README content
The versions of the GitHub actions 'checkout' and 'setup-python' were updated from 'v2' to 'v4' in the python-app.yml file to use the latest features and improvements.

In the README.md file, extensive content about the project, its tech stack, how to set it up, and other project-related information was added. This provides comprehensive details about the project for potential contributors and users.